# openapi.yaml
openapi: 3.0.3
info:
  title: AlfTekPro Multi-Tenant HRMS API
  version: 1.0.0
  description: |
    Multi-tenant HRMS system with row-level isolation, supporting multiple regions (UAE, USA, India).

    ## Architecture
    - **Authentication**: JWT Bearer tokens with tenant_id in claims
    - **Multi-Tenancy**: Row-level isolation with automatic query filtering
    - **Regions**: UAE (RTL/Arabic), USA (LTR/English), India (LTR/Hindi)

    ## Key Features
    - Employee Management with Temporal Job History
    - Shift & Roster Management
    - Attendance Tracking with Geofencing
    - Leave Management with Approval Workflows
    - Unified Action Center for Approvals
    - Payroll Processing with Regional Logic
    - Asset Management
    - Dynamic Forms (Region-Specific Fields)

  contact:
    name: AlfTekPro Support
    email: support@alftekpro.com
  license:
    name: Proprietary
servers:
  - url: https://api.alftekpro.com/v1
    description: Production Server
  - url: https://dev-api.alftekpro.com/v1
    description: Development Server
  - url: http://localhost:5001/v1
    description: Local Development
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticate user and return JWT token with tenant context
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Generate new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with claims: tenant_id, user_id, role

        Example header: Authorization: Bearer <token>
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Bad Request - Validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errorCode:
          type: string
          description: i18n error code
        errors:
          type: object
          additionalProperties:
            type: string
          description: Validation errors keyed by field name

    # ==================== AUTH SCHEMAS ====================
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
              description: JWT access token
            refreshToken:
              type: string
            expiresAt:
              type: string
              format: date-time
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                email:
                  type: string
                name:
                  type: string
                role:
                  type: string
                  enum: [SA, TA, MGR, PA, EMP]
                tenantId:
                  type: string
                  format: uuid
                  nullable: true
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    RefreshTokenResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            expiresAt:
              type: string
              format: date-time

    # ==================== PLATFORM SCHEMAS ====================
...